+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.15.0
 
 * Adding `clustering/inverted-index` namespace.
+* Adding the `asymmetric` option to `clustering/naive`.
 * Moving `stats/ngrams` to `tokenizers/ngrams`.
 * Dropping the `stats/tfidf` namespace.